+++ src/his_solver.h
@@ -7,28 +7,28 @@
 
 class SegmentTree {
  public:
-  SegmentTree(int n) : n_(n) { tree_ = std::vector<long long>(n_ + n_, 0); }
+  SegmentTree(int n) : n_(n) { tree_ = std::vector<int64_t>(n_ + n_, 0); }
 
   // Updates value of node at pos to v, if v is greater than the current value.
-  void update(int pos, long long v);
+  void update(int pos, int64_t v);
 
   // Returns max value in range [l, r).
-  long long query(int l, int r);
+  int64_t query(int l, int r);
 
  private:
   const int n_;
-  std::vector<long long> tree_;
+  std::vector<int64_t> tree_;
 };
 
 // Solves his2D problem using Divide & Conquer approach together with Dynamic
 // Programming and Segment Tree.
 class His2DSolver {
  public:
-  His2DSolver(std::vector<std::pair<long long, long long>> v,
+  His2DSolver(std::vector<std::pair<int64_t, int64_t>> v,
               std::vector<int> weight, const int epsilon);
 
   // Returns weight of HIS.
-  long long Weight() { return *std::max_element(dp_.begin(), dp_.end()); }
+  int64_t Weight() { return *std::max_element(dp_.begin(), dp_.end()); }
 
   // Returns positions of HIS elements.
   std::vector<int> His();
@@ -40,9 +40,9 @@ class His2DSolver {
 
   std::vector<std::pair<int, bool>> Sort(int l, int r);
 
-  std::vector<std::pair<long long, long long>> v_;
+  std::vector<std::pair<int64_t, int64_t>> v_;
   std::vector<int> weight_;
-  std::vector<long long> dp_;
+  std::vector<int64_t> dp_;
   int epsilon_;
 };
